--- src/components/welcome/styles.scss
@@ -1,32 +1,0 @@
-@use '@/static/styles/main.scss';
-
-.welcome {
-  height: 100%;
-  &__form {
-    display: flex;
-    justify-content: flex-end;
-    flex-direction: column;
-    align-content: flex-end;
-    align-items: center;
-    color: var(--text-color);
-  }
-  &__button {
-    height: 50px;
-    width: 100px;
-    font-size: 15px;
-    background-color: var(--gray-color);
-    border-radius: 15px;
-    cursor: pointer;
-    color: var(--text-color);
-  }
-}
-
-.dark-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5); // Полупрозрачный черный фон
-  z-index: 999; // Убедитесь, что он выше других элементов
-}
