+++ src/static/styles/main.scss
@@ -15,6 +15,7 @@ body {
   background-color: $body-background-color;
   margin: 0;
   height: 100vh;
+  font-family: 'Roboto', sans-serif;
 }
 body::-webkit-scrollbar {
   display: none;
